# doc/yaml/api.yaml
openapi: 3.0.3
info:
  title: "Squealer Backend"
  description: "Backend API for all Squealer Products"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
...
  /channels/:
    get:
      tags:
        - Channel
      operationId: getChannels
      parameters:
        - name: page
          in: query
          description: To paginate results. A non-positive value disables pagination.
          schema:
            type: integer
            example: 3
        - name: results_per_page
          in: query
          description: Number of results to return in each query.
          schema:
            type: integer
            default: 100 
            minimum: 1
            example: 50
        - name: owner
          in: query
          description: Filter by owner's Handle. If no channel owned by a user with handle is found it will return an empty array
          schema:
            type: string
            example: owner333
        - name: publicChannel  
          in: query
          description: Only get channels that are public (if the field is true) or private (if it's false)
          schema:
            type: boolean
            example: False
        - name: name  
          in: query
          description: Only get channel whose names contain the given string. Matching is case-insensitive.
          schema:
            type: string
            example: test-channel
        - name: member 
          in: query
          description: Only get channel that the given user is a member of.
          schema:
            type: string
            description: The user's handle
            example: someHandle333
        - name: owner 
          in: query
          description: Only get channels created by the given user.
          schema:
            type: string
            description: The user's handle
            example: someHandle333
        - name: member 
          in: query
          description: Only get channels the user is a member of.
          schema:
            type: string
            description: The user's handle
            example: someHandle333
        - name: official 
          in: query
          description: Only get official/non official channels.
          schema:
            type: boolean
            example: True
        - name: namesOnly 
          in: query
          description: If true the endpoint will only return the channel's names.
          schema:
            type: boolean
            example: True
      description: Get all channels.
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: array
                items:
                  anyOf:
                    - $ref: '#/components/schemas/Channel'
                    - type: array
                      items: 
                        type: string 
                        description: The channel names.
                      example: ['channel1', 'channel2']
        '401':
          description: Unauthorized operation
  /channels/{name}:
    post:
      tags:
        - Channel
      description: Create a new channel with the given name. Requesting user will be the creator.
      operationId: createChannel
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: name
          in: path
          description: The Channel's name
          required: True
          example: channelAboutMountainPictures88
          schema:
            type: string
            minLength: 6
      requestBody:
        required: True
        description: The new channel's data.
        content:
          application/json: 
            schema:
              type: object
              properties:
                description: 
                  type: string 
                  description: The channel's description
                publicChannel:
                  type: boolean
                  example: True 
                  default: True 
                  description: Wether to make the channel public or not. 
                official: 
                  type: boolean
                  example: True 
                  default: False 
                  description: Wether to make the channel official or not. Only admins can create official channels.
              required: [publicChannel, official]
      responses:
        '200':
          description: Channel Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '401':
          description: Unauthorized operation.
        '409':
          description: Channel name already taken.
    get:
      tags:
        - Channel
      description: Create a new channel with the given name.
      operationId: getChannel
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: name
          in: path
          description: The Channel's name
          required: True
          example: channelAboutMountainPictures88
          schema:
            type: string
            minLength: 6
      responses:
        '200':
          description: Request Successful 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '401':
          description: Unauthorized operation.
        '409':
          description: Channel not found.
    put:
      tags:
        - Channel
      description: Change channel data, must be the owner.
      operationId: writeChannel
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: name
          in: path
          description: The new Channel's name
          required: True
          example: channelAboutMountainPictures88
          schema:
            type: string
            minLength: 6
      requestBody:
        required: True
        description: The data to be changed.
        content:
          application/json: 
            schema:
              type: object
              properties:
                owner:
                  type: string
                  description: The new owner's handle
                  example: ownerHandle777
                newName:
                  type: string
                  description: The new channel's Name
                  example: Channel34
                publicChannel:
                  type: boolean
                  description: The new channel's private field, its called like that because javascript.
                  example: true
                description:
                  type: string
                  description: The channel's description
                  example: This is a channel about how javascript is a great programming language and not     at all a meme
      responses:
        '200':
          description: Channel Changed Successfully
        '401':
          description: Unauthorized operation.
        '409':
          description: No channel with the given name.
    delete:
      tags:
        - Channel
      description: Delete given channel. Must be the creator.
      operationId: deleteChannel
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: name
          in: path
          required: True
          description: The channel's name.
          schema:
            type: string
            example: awesomeChannel9000
      responses:
        '200': 
          description: Channel Deleted.
        '401':
          description: Unauthorized operation.
        '409':
          description: No channel with the given name.
  /messages/channel/{name}/:
    get:
      description: Get messages addressed to a given channel. Must have reading permission.
      tags:
          - Channel
      operationId: getChannelMessages
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: name
          in: path
          required: True
          description: The channel's name.
          schema:
            type: string
            example: awesomeChannel9000
        - name: page
          in: query
          description: Paginate results. A non-positive value disables pagination.
          schema:
            type: integer
            default: 1
            example: 3
        - name: results_per_page
          in: query
          description: Number of results to return in each query.
          schema:
            type: integer
            default: 100 
            minimum: 1
            example: 50
      responses:
        '200': 
          description: Request Successful.
          content:
            application/json:
              schema:
                type: array 
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized operation.
        '409':
          description: No channel with the given name.
    delete:
      description: Remove channel from the dest of all messages.
      tags:
          - Channel
      operationId: deleteChannelMessages
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: name
          in: path
          required: True
          description: The channel's name.
          schema:
            type: string
            example: awesomeChannel9000
      responses:
        '200': 
          description: Request Successful.
        '401':
          description: Unauthorized operation.
        '409':
          description: No channel with the given name.
  /channels/{name}/available:
    get:
      tags:
        - Channel
      description: Check wether the given name is available or not.
      operationId: availableChannel
      parameters:
        - name: name
          required: True
          in: path
          schema:
            type: string
            example: canIUseThisName99
      responses:
        '200':
          description: Successfully checked, answer is in body
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
                    description: true if the name is available, false otherwise.
                required: [available]
        '409':
          description: Did not provide a name
        '500':
          description: Internal server error.
  /channel/{name}/members: 
    post: 
      tags:
        - Channel
        - User
      description: Add and remove members.
      operationId: writeChannelMembers
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: name
          in: path
          description: The Channel's name
          required: True
          example: channelAboutMountainPictures88
          schema:
            type: string
            minLength: 6
      requestBody:
        required: True
        description: The data to be changed.
        content:
          application/json: 
            schema:
              type: object
              properties:
                addMembers:
                  type: array
                  description: Array of users's handles to be added as members.
                  items:
                    type: string 
                    example: handle1123
                removeMembers:
                  type: array
                  description: Array of handles to remove from members. 
                    Users who are also editors will be removed from the editors as well.
                  items:
                    type: string 
                    example: handle1123
                removeRequests:
                  type: array
                  description: Array of handles to remove from member requests.
                  items:
                    type: string 
                    example: handle1123
      responses:
        '200': 
          description: Operation successful.
        '401':
          description: Unauthorized operation.
        '409':
          description: No channel with the given name.
  /channel/{name}/editors: 
    post: 
      tags:
        - Channel
        - User
      description: Add and remove editors.
      operationId: writeChannelEditors
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: name
          in: path
          description: The Channel's name
          required: True
          example: channelAboutMountainPictures88
          schema:
            type: string
            minLength: 6
      requestBody:
        required: True
        description: The data to be changed.
        content:
          application/json: 
            schema:
              type: object
              properties:
                addEditors:
                  type: array
                  description: Array of users's handles to be added as editors. Users who are 
                    not members will also be added to the members list.
                  items:
                    type: string 
                    example: handle1123
                removeEditors:
                  type: array
                  description: Array of handles to remove from the editor list. 
                  items:
                    type: string 
                    example: handle1123
                removeRequests:
                  type: array
                  description: Array of handles to remove from editor requests.
                  items:
                    type: string 
                    example: handle1123
      responses:
        '200': 
          description: Operation successful.
        '401':
          description: Unauthorized operation.
        '409':
          description: No channel with the given name.
...
  /public/channels/:
    get:
      tags:
        - Channel
        - Public
      operationId: publicGetChannels
      parameters:
        - name: page
          in: query
          description: To paginate results. A non-positive value disables pagination.
          schema:
            type: integer
            example: 3
        - name: results_per_page
          in: query
          description: Number of results to return in each query.
          schema:
            type: integer
            default: 100 
            minimum: 1
            example: 50
        - name: owner
          in: query
          description: Filter by owner's Handle. If no channel owned by a user with handle is found it will return an empty array
          schema:
            type: string
            example: owner333
        - name: name  
          in: query
          description: Only get channel whose names contain the given string. Matching is case-insensitive.
          schema:
            type: string
            example: test-channel
        - name: member 
          in: query
          description: Only get channel that the given user is a member of.
          schema:
            type: string
            description: The user's handle
            example: someHandle333
        - name: owner 
          in: query
          description: Only get channels created by the given user.
          schema:
            type: string
            description: The user's handle
            example: someHandle333
        - name: member 
          in: query
          description: Only get channels the user is a member of.
          schema:
            type: string
            description: The user's handle
            example: someHandle333
        - name: official 
          in: query
          description: Only get official/non official channels.
          schema:
            type: boolean
            example: True
        - name: namesOnly 
          in: query
          description: If true the endpoint will only return the channel's names.
          schema:
            type: boolean
            example: True
      description: Get all official channels.
      responses:
        '200':
          description: Valid request
          content:
            application/json:
              schema:
                type: array
                items:
                  anyOf:
                    - $ref: '#/components/schemas/Channel'
                    - type: array
                      items: 
                        type: string 
                        description: The channel names.
                      example: ['channel1', 'channel2']
        '401':
          description: Unauthorized operation
...
components:
  schemas:
...
    Message:
      allOf:
        - $ref: '#/components/schemas/MessageInfo'
        - $ref: '#/components/schemas/MessageImpressions'
        - type: object
          required: [id, meta]
          properties:
            id:
              type: integer
              format: int64
              description: The message's numerical id
            meta:
              $ref: '#/components/schemas/MessageMeta'
    MessageImpressions:
      type: object
      required: [viewed, reactions]
      properties:
        viewed:
              type: integer
              description: Number of time the message has been viewed
              minimum: 0
        reactions:
          type: object
          required: [positive, negative]
          description: Bookkeeping data on reactions
          properties:
            positive:
              type: object
              description: Numbers of positive reactions divided by types
              properties:
                agree:
                  type: integer
                  minimum: 0
                strongAgree: 
                  type: integer
                  minimum: 0
            negative:
              type: object
              description: Numbers of negative reactions divided by types
              properties:
                disagree:
                  type: integer
                  minimum: 0
                strongDisagree: 
                  type: integer
                  minimum: 0
    MessageInfo:
      type: object
      required: [content, posted, author, dest]
      properties:
        content:
          type: object
          properties:
            text: 
              type: string
              example: This is an example squeal that contains very interesting info.
            image:
              type: string
              format: binary
              description: An image
        posted:
          type: string
          description: Message's timestamp
          format: date-time
          example: "2021-01-30T08:30:00Z"
        author:
          type: string
          description: The author's handle
          example: pieraldo196
        dest:
          type: array
          items:
            type: string
            description: The handle for single individuals, the channel name for channels
          example: ['@gianni', '@alberto', '§canalebello']
    MessageMeta:
      type: object 
      properties:
        created:
          type: string 
          format: date 
          description: Message creation date 
        lastModified:
          type: string 
          format: date 
          description: Last time the message was modified
        geo:
          $ref: '#/components/schemas/GeoSchema'
        impressions:
          type: number 
          description: Number of times the message was viewed.
          minimum: 0
    GeoSchema:
      type: object 
      properties:
        type:
          type: string
          enum: 
            - Point
          default: Point 
        coordinates:
          type: array 
          items: 
            type: number
          maxLength: 2
          minLength: 2
          example: [45.89, 42.069]
    Channel:
      type: object
      required: [name, created, owner, publicChannel, official]
      properties:
        name: 
          type: string
          example: AwesomeChannel42
        created:
          type: string
          format: date-time
          description: Channel's creation timestamp
          example: "2021-01-30T08:30:00Z"
        creator:
          type: string
          description: Owner's Handle
          example: channel3Owner
        description: 
          type: string
          description: A brief description of the channel
          example: This is a channel dedicated to pictures of fluffy kittens
        members:
          type: array
          items:
            type: string
            description: Handles of the users with reading permission. The creator is always a member.
          minLength: 1       
        editors:
          type: array
          items:
            type: string
            description: Handles of the users with reading and writing permission. The creator is always an editor.
          minLength: 1
        memberRequests:
          type: array
          items:
            type: string
            description: Handles of the users requesting reading permissions.
          minLength: 0      
        editorRequests:
          type: array
          items:
            type: string
            description: Handles of the users requesting reading and writing permission. 
              If a user is added to the editors they will also be automatically added to the members.
          minLength: 0
        publicChannel:
          type: boolean
          default: true 
          description: If true the channel member/editor and member/editor requests can be viewed by anyone.
        official:
          type: boolean 
          default: false 
          description: Wether the channel is an official squealer channel or not.
  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Requires Requires a JWT token generated with user credentials. Even with a correctly formed JWT, a user can only alter their own account.
    proAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT
        description: Requires a JWT token generated with pro user credentials.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Requires a JWT token generated with admin credentials.      